requirements.txt
django-widget-tweaks
djangorestframework
pytz
django-simple-math-captcha